LPS-64238 move
