+++ modules/apps/foundation/portal/portal-dao-orm-custom-sql/bnd.bnd
@@ -1,7 +1,7 @@
-Bundle-Name: Liferay DAO ORM Custom
-Bundle-SymbolicName: com.liferay.portal.dao.orm
+Bundle-Name: Liferay DAO ORM Custom SQL
+Bundle-SymbolicName: com.liferay.portal.dao.orm.custom.sql
 Bundle-Version: 1.0.1
-Export-Package: com.liferay.portal.dao.orm
+Export-Package: com.liferay.portal.dao.orm.custom.sql
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title:
 -include: ../../app.bnd
