+++ modules/apps/foundation/portal/portal-dao-orm-custom-sql/src/main/java/com/liferay/portal/dao/orm/custom/sql/CustomSQLUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.dao.orm;
+package com.liferay.portal.dao.orm.custom.sql;
 
 import com.liferay.portal.kernel.dao.orm.QueryDefinition;
 import com.liferay.portal.kernel.dao.orm.WildcardMode;
